Make a method to reconstruct K, and add mock code in it's own source file
