+++ pkg/wallet/wallet_test.go
@@ -5,15 +5,11 @@ import (
 	"math/big"
 	"os"
 	"testing"
-	"time"
 
 	"github.com/dusk-network/dusk-blockchain/pkg/core/block"
 	"github.com/dusk-network/dusk-blockchain/pkg/wallet/database"
 	"github.com/dusk-network/dusk-blockchain/pkg/wallet/transactions"
-	"github.com/dusk-network/dusk-crypto/mlsag"
 	"github.com/dusk-network/dusk-wallet/key"
-
-	"math/rand"
 
 	"github.com/bwesterb/go-ristretto"
 	"github.com/stretchr/testify/assert"
@@ -28,15 +24,15 @@ func TestNewWallet(t *testing.T) {
 	assert.Nil(t, err)
 	defer os.RemoveAll(dbPath)
 
-	w, err := New(randReader, netPrefix, db, generateDecoys, fetchInputs, "pass")
+	w, err := New(randReader, netPrefix, db, GenerateDecoys, GenerateInputs, "pass")
 	assert.Nil(t, err)
 
 	// wrong wallet password
-	loadedWallet, err := LoadFromFile(netPrefix, db, generateDecoys, fetchInputs, "wrongPass")
+	loadedWallet, err := LoadFromFile(netPrefix, db, GenerateDecoys, GenerateInputs, "wrongPass")
 	assert.NotNil(t, err)
 
 	// correct wallet password
-	loadedWallet, err = LoadFromFile(netPrefix, db, generateDecoys, fetchInputs, "pass")
+	loadedWallet, err = LoadFromFile(netPrefix, db, GenerateDecoys, GenerateInputs, "pass")
 	assert.Nil(t, err)
 
 	assert.Equal(t, w.PublicKey(), loadedWallet.PublicKey())
@@ -55,7 +51,7 @@ func TestReceivedTx(t *testing.T) {
 	assert.Nil(t, err)
 	defer os.RemoveAll(dbPath)
 
-	w, err := New(randReader, netPrefix, db, generateDecoys, fetchInputs, "pass")
+	w, err := New(randReader, netPrefix, db, GenerateDecoys, GenerateInputs, "pass")
 	assert.Nil(t, err)
 
 	tx, err := w.NewStandardTx(fee)
@@ -116,7 +112,7 @@ func generateWallet(t *testing.T, netPrefix byte, path string) *Wallet {
 	assert.Nil(t, err)
 	defer os.RemoveAll(path)
 
-	w, err := New(randReader, netPrefix, db, generateDecoys, fetchInputs, "pass")
+	w, err := New(randReader, netPrefix, db, GenerateDecoys, GenerateInputs, "pass")
 	assert.Nil(t, err)
 	return w
 }
@@ -137,61 +133,10 @@ func generateStandardTx(t *testing.T, receiver key.PublicAddress, amount int64, 
 	return tx
 }
 
-func fetchInputs(netPrefix byte, db *database.DB, totalAmount int64, key *key.Key) ([]*transactions.Input, int64, error) {
-
-	// This function shoud store the inputs in a database
-	// Upon calling fetchInputs, we use the keyPair to get the privateKey from the
-	// one time pubkey
-
-	var inputs []*transactions.Input
-	numInputs := 7
-
-	addresses, R := generateOutputAddress(key, netPrefix, numInputs)
-
-	rand.Seed(time.Now().Unix())
-	randAmount := rand.Int63n(totalAmount) + int64(totalAmount)/int64(numInputs) + 1 // [totalAmount/4 + 1, totalAmount*4]
-	remainder := (randAmount * int64(numInputs)) - totalAmount
-	if remainder < 0 {
-		remainder = 0
-	}
-
-	for index, addr := range addresses {
-		var amount, mask ristretto.Scalar
-		amount.SetBigInt(big.NewInt(randAmount))
-		mask.Rand()
-
-		// Fetch the privKey for each addresses
-		privKey, _ := key.DidReceiveTx(R, *addr, uint32(index))
-
-		input := transactions.NewInput(amount, mask, *privKey)
-		inputs = append(inputs, input)
-	}
-
-	return inputs, remainder, nil
-}
-
 func generateSendAddr(t *testing.T, netPrefix byte, randKeyPair *key.Key) key.PublicAddress {
 	pubAddr, err := randKeyPair.PublicKey().PublicAddress(netPrefix)
 	assert.Nil(t, err)
 	return *pubAddr
-}
-
-func generateOutputAddress(keyPair *key.Key, netPrefix byte, num int) ([]*key.StealthAddress, ristretto.Point) {
-	var res []*key.StealthAddress
-
-	sendersPubKey := keyPair.PublicKey()
-
-	var r ristretto.Scalar
-	r.Rand()
-
-	var R ristretto.Point
-	R.ScalarMultBase(&r)
-
-	for i := 0; i < num; i++ {
-		stealthAddr := sendersPubKey.StealthAddress(r, uint32(i))
-		res = append(res, stealthAddr)
-	}
-	return res, R
 }
 
 // https://www.dotnetperls.com/duplicates-go
@@ -216,29 +161,6 @@ func sliceToPoint(t *testing.T, b []byte) ristretto.Point {
 	copy(byts[:], b)
 	c.SetBytes(&byts)
 	return c
-}
-
-func generateDualKey() mlsag.PubKeys {
-	pubkeys := mlsag.PubKeys{}
-
-	var primaryKey ristretto.Point
-	primaryKey.Rand()
-	pubkeys.AddPubKey(primaryKey)
-
-	var secondaryKey ristretto.Point
-	secondaryKey.Rand()
-	pubkeys.AddPubKey(secondaryKey)
-
-	return pubkeys
-}
-
-func generateDecoys(numMixins int) []mlsag.PubKeys {
-	var pubKeys []mlsag.PubKeys
-	for i := 0; i < numMixins; i++ {
-		pubKeyVector := generateDualKey()
-		pubKeys = append(pubKeys, pubKeyVector)
-	}
-	return pubKeys
 }
 
 func randReader(b []byte) (n int, err error) {
